requirements-dev.txt
coverage==3.7.1
nose==1.3.7
rednose==0.4.3
pyhamcrest==1.8.5
requests-mock==0.6.0
